feat: switch to using a cache for results
- create a results cache and loader to process the results file into the
  cache
- simplify the treeview provider to remove the loading of results and
  handle each results file as it is processed
- extract the problem updating into its own singleton with the
  diagnostics state updated on each run
